+++ @remirror/extension-code-block/src/patches.d.ts
@@ -1,4 +1,4 @@
-declare module 'react-syntax-highlighter/prism-light' {
+declare module 'react-syntax-highlighter/dist/cjs/prism-light' {
   import * as React from 'react';
   import { SyntaxHighlighterProps } from 'react-syntax-highlighter';
   export default class SyntaxHighlighter extends React.Component<SyntaxHighlighterProps> {